prop: cadical: Reorganize CaDiCaL wrapper code.
Moves implementations of IPASIRUP propagator and proof tracer into
separate files.
